requirements.txt
requests>=2.32.5,<3
rich>=14.3.3,<15
prettytable>=3.17.0,<4
click>=8.0.0,<9
json-repair>=0.58.6,<1
pathvalidate>=3.3.1,<4
platformdirs>=4.9.4,<5
emoji>=2.15.0,<3
bleach>=6.3.0,<7
beautifulsoup4>=4.14.3,<5
parsel>=1.11.0,<2
tqdm>=4.67.3,<5
pycryptodomex>=3.23.0,<4
pycryptodome>=3.23.0,<4
cryptography>=46.0.5,<47
tldextract>=5.3.1,<6
fake-useragent>=2.2.0,<3
nodejs-wheel>=24.14.0,<25
pyfreeproxy>=0.4.0,<1
DrissionPage>=4.1.1.2,<5
questionary>=2.1.1,<3
pywidevine>=1.9.0,<2
curl-cffi>=0.14.0,<1
m3u8>=6.0.0,<7
brotli>=1.0.9,<2
lxml>=6.0.2,<7
filetype>=1.2.0,<2
puremagic>=2.1.1,<3; python_version >= "3.12"
puremagic>=1.30,<2; python_version < "3.12"